More modest module monikers
