+++ glamod_marine_processing/qc_suite/modules/buoy_tracking_qc.py
@@ -13,7 +13,7 @@ import numpy as np
 
 from .astronomical_geometry import sunangle
 from .auxiliary import failed, isvalid, passed, untestable, untested
-from .next_level_track_check_qc import do_iquam_track_check, inspect_arrays
+from .qc_sequential_reports import do_iquam_track_check, inspect_arrays
 from .spherical_geometry import sphere_distance
 from .statistics import trim_mean, trim_std
 from .time_control import convert_date_to_hours, dayinyear
